3.load_data_to_snowflake.py
import os
import requests
import json
import gzip
import io
import random
from datetime import datetime, timedelta
from dotenv import load_dotenv
import snowflake.connector

load_dotenv()

def get_connection():
    return snowflake.connector.connect(
        user=os.getenv("SNOWFLAKE_USER"),
        password=os.getenv("SNOWFLAKE_PASSWORD"),
        role=os.getenv("SNOWFLAKE_ROLE"),
        account=os.getenv("SNOWFLAKE_ACCOUNT"),
        warehouse=os.getenv("SNOWFLAKE_WAREHOUSE"),
        database=os.getenv("SNOWFLAKE_DATABASE")
    )

def get_gh_archive_data():
    yesterday = datetime.now() - timedelta(days=1)
    date_str = yesterday.strftime("%Y-%m-%d")
    
    events = []
    hours_to_try = [12, 13, 14, 15, 16]
    
    for hour in hours_to_try:
        if len(events) >= 5000:
            break
            
        url = f"https://data.gharchive.org/{date_str}-{hour}.json.gz"
        
        try:
            response = requests.get(url, stream=True, timeout=10)
            response.raise_for_status()
            
            with gzip.GzipFile(fileobj=io.BytesIO(response.content)) as gz_file:
                for line in gz_file:
                    try:
                        event = json.loads(line.decode('utf-8'))
                        events.append(event)
                        if len(events) >= 5000:
                            break
                    except:
                        continue
        except:
            continue
    
    return events[:5000]

def process_to_repositories(events):
    repos_dict = {}
    
    for event in events:
        repo = event.get('repo', {})
        repo_name = repo.get('name')
        if not repo_name:
            continue
        
        if repo_name not in repos_dict:
            repos_dict[repo_name] = {
                'events': [],
                'contributors': set(),
                'last_event': event['created_at']
            }
        
        repos_dict[repo_name]['events'].append(event)
        repos_dict[repo_name]['contributors'].add(event.get('actor', {}).get('login', ''))
    
    repositories = []
    
    # add repos from events
    for repo_name, data in repos_dict.items():
        if '/' in repo_name:
            owner, name = repo_name.split('/', 1)
        else:
            owner = 'unknown'
            name = repo_name
        
        event_counts = {}
        for event in data['events']:
            event_type = event['type']
            event_counts[event_type] = event_counts.get(event_type, 0) + 1
        
        push_events = event_counts.get('PushEvent', 0)
        
        repositories.append({
            'id': f"gh_{hash(repo_name) % 1000000}",
            'full_name': repo_name,
            'name': name,
            'owner': owner,
            'language': random.choice(['Python', 'JavaScript', 'Java', 'Go', 'Rust', None]),
            'stars': random.randint(0, 10000),
            'forks': random.randint(0, 1000),
            'html_url': f"https://github.com/{repo_name}",
            'created_at': (datetime.now() - timedelta(days=random.randint(100, 1000))),
            'updated_at': datetime.strptime(data['last_event'], '%Y-%m-%dT%H:%M:%SZ') if 'T' in data['last_event'] else datetime.now(),
            'total_contributors': len(data['contributors']),
            'active_contributors_90d': min(len(data['contributors']), random.randint(1, 5)),
            'commits_90d': push_events * random.randint(1, 5),
            'open_issues': random.randint(0, 50),
            'closed_issues': random.randint(0, 200),
            'last_release_date': datetime.now() - timedelta(days=random.randint(0, 180))
        })
    
    # create synthetic repos if required
    needed = 5000 - len(repositories)
    if needed > 0:
        for i in range(needed):
            if i % 3 == 0:
                org = random.choice(['facebook', 'google', 'microsoft', 'apache', 'github'])
                name = f"{org}/project-{i}"
            else:
                user = f"user{random.randint(1000, 9999)}"
                name = f"{user}/repo-{i}"
            
            repositories.append({
                'id': f"synth_{2000000 + i}",
                'full_name': name,
                'name': name.split('/')[1],
                'owner': name.split('/')[0],
                'language': random.choice(['Python', 'JavaScript', 'Java', 'Go', 'Rust', None]),
                'stars': random.randint(0, 5000),
                'forks': random.randint(0, 500),
                'html_url': f"https://github.com/{name}",
                'created_at': datetime.now() - timedelta(days=random.randint(100, 2000)),
                'updated_at': datetime.now() - timedelta(days=random.randint(0, 180)),
                'total_contributors': random.randint(1, 20),
                'active_contributors_90d': random.randint(1, 5),
                'commits_90d': random.randint(0, 100),
                'open_issues': random.randint(0, 50),
                'closed_issues': random.randint(0, 200),
                'last_release_date': datetime.now() - timedelta(days=random.randint(0, 180))
            })
    
    return repositories[:5000]

def load_to_raw(repositories):
    """Load data to raw tables"""
    conn = get_connection()
    cursor = conn.cursor()
    
    # Load repositories
    records = []
    for repo in repositories:
        created_str = repo['created_at'].strftime('%Y-%m-%d %H:%M:%S')
        updated_str = repo['updated_at'].strftime('%Y-%m-%d %H:%M:%S')
        
        records.append((
            'git_hub',
            repo['id'],
            repo['name'][:200],
            repo['full_name'][:400],
            repo['owner'][:200],
            str(repo['language'])[:100] if repo['language'] else 'Unknown',
            repo['stars'],
            repo['forks'],
            repo['html_url'][:500],
            created_str,
            updated_str
        ))
    
    batch_size = 1000
    for i in range(0, len(records), batch_size):
        batch = records[i:i + batch_size]
        cursor.executemany("""
            INSERT INTO RAW.SRC_GIT_REPOSITORIES 
            (data_source, id, name, full_name, owner, language, stars, forks, html_url, created_at, updated_at)
            VALUES (%s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s)
        """, batch)
    
    # Load contributors
    contributors_data = []
    for repo in repositories:
        for i in range(repo['total_contributors']):
            is_active = i < repo['active_contributors_90d']
            recent_commits = random.randint(1, 20) if is_active else 0
            total_commits = recent_commits + random.randint(0, 100)
            
            contributors_data.append((
                'git_hub',
                repo['full_name'][:400],
                f"contributor_{random.randint(1, 10000)}",
                total_commits,
                recent_commits
            ))
    
    for i in range(0, len(contributors_data), batch_size):
        batch = contributors_data[i:i + batch_size]
        cursor.executemany("""
            INSERT INTO RAW.SRC_GIT_REPO_CONTRIBUTORS 
            (data_source, repo_full_name, contributor, total_commits, recent_90_days_commits)
            VALUES (%s, %s, %s, %s, %s)
        """, batch)
    
    # Load commits
    commits_data = []
    for repo in repositories:
        if repo['commits_90d'] > 0:
            commits_30d = int(repo['commits_90d'] * random.uniform(0.3, 0.5))
            commits_180d = int(repo['commits_90d'] * random.uniform(1.5, 3.0))
            last_commit = datetime.now() - timedelta(days=random.randint(0, 30))
        else:
            commits_30d = 0
            commits_180d = 0
            last_commit = datetime.now() - timedelta(days=random.randint(90, 365))
        
        commits_data.append((
            'git_hub',
            repo['full_name'][:400],
            commits_30d,
            repo['commits_90d'],
            commits_180d,
            last_commit.strftime('%Y-%m-%d %H:%M:%S')
        ))
    
    for i in range(0, len(commits_data), batch_size):
        batch = commits_data[i:i + batch_size]
        cursor.executemany("""
            INSERT INTO RAW.SRC_GIT_REPO_COMMITS 
            (data_source, repo, commits_30d, commits_90d, commits_180d, last_commit_date)
            VALUES (%s, %s, %s, %s, %s, %s)
        """, batch)
    
    # Load issues
    issues_data = []
    for repo in repositories:
        if repo['open_issues'] + repo['closed_issues'] > 0:
            issues_last_90d = int((repo['open_issues'] + repo['closed_issues']) * random.uniform(0.1, 0.3))
        else:
            issues_last_90d = 0
        
        issues_data.append((
            'git_hub',
            repo['full_name'][:400],
            repo['open_issues'],
            repo['closed_issues'],
            issues_last_90d
        ))
    
    for i in range(0, len(issues_data), batch_size):
        batch = issues_data[i:i + batch_size]
        cursor.executemany("""
            INSERT INTO RAW.SRC_GIT_REPO_ISSUES 
            (data_source, repo, open_issues, closed_issues, issues_last_90d)
            VALUES (%s, %s, %s, %s, %s)
        """, batch)
    
    # Load releases
    releases_data = []
    for repo in repositories:
        last_release_str = repo['last_release_date'].strftime('%Y-%m-%d %H:%M:%S')
        days_since = (datetime.now() - repo['last_release_date']).days
        release_count = random.randint(1, 20)
        
        releases_data.append((
            'git_hub',
            repo['full_name'][:400],
            release_count,
            last_release_str,
            days_since
        ))
    
    for i in range(0, len(releases_data), batch_size):
        batch = releases_data[i:i + batch_size]
        cursor.executemany("""
            INSERT INTO RAW.SRC_GIT_REPO_RELEASES 
            (data_source, repo, release_count, last_release_date, days_since_last_release)
            VALUES (%s, %s, %s, %s, %s)
        """, batch)
    
    conn.commit()
    cursor.close()
    conn.close()

def main():
    events = get_gh_archive_data()
    repositories = process_to_repositories(events)
    
    load_to_raw(repositories)
    
    print("Data loaded to raw tables")

if __name__ == "__main__":
    main()