requirements.txt

snowflake-connector-python==3.12.0
python-dotenv==1.0.1
requests==2.32.3
pandas==2.2.3